requirements.txt
Flask==2.0.3
Flask-Babel==2.0.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
ReportLab==3.6.8
openai==0.7.0
python-dotenv==0.19.2
